+++ internal/backends/linuxkms/display/mod.rs
@@ -1,0 +1,86 @@
+// Copyright © SixtyFPS GmbH <info@slint.dev>
+// SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-Royalty-free-1.1 OR LicenseRef-Slint-commercial
+
+use i_slint_core::api::PhysicalSize;
+
+pub trait Presenter {
+    // Present updated front-buffer to the screen
+    fn present(&self) -> Result<(), Box<dyn std::error::Error + Send + Sync>>;
+}
+
+#[cfg(any(feature = "renderer-skia-opengl", feature = "renderer-femtovg"))]
+pub mod egldisplay;
+#[cfg(feature = "renderer-skia-vulkan")]
+pub mod vulkandisplay;
+
+/// This enum describes the way the output is supposed to be rotated to simulate
+/// a screen rotation. This is implemented entirely inside the actual renderer.
+#[non_exhaustive]
+#[derive(Default, Copy, Clone, Eq, PartialEq, Debug)]
+pub enum SyntheticDisplayRotation {
+    /// No rotation
+    #[default]
+    NoRotation,
+    /// Rotate 90° to the left
+    Rotate90,
+    /// 180° rotation (upside-down)
+    Rotate180,
+    /// Rotate 90° to the right
+    Rotate270,
+}
+
+impl TryFrom<&str> for SyntheticDisplayRotation {
+    type Error = String;
+
+    fn try_from(value: &str) -> Result<Self, Self::Error> {
+        let angle: usize = value.parse().map_err(|_| {
+            format!("Invalid value for rotation. Must be unsigned integral, found {value}")
+        })?;
+        Ok(match angle {
+            0 => Self::NoRotation,
+            90 => Self::Rotate90,
+            180 => Self::Rotate180,
+            270 => Self::Rotate270,
+            _ => {
+                return Err(format!(
+                    "Invalid value for rotation. Must be one of 0, 90, 180, or 270"
+                ))
+            }
+        })
+    }
+}
+
+impl SyntheticDisplayRotation {
+    pub fn screen_size_to_rotated_window_size(&self, screen_size: PhysicalSize) -> PhysicalSize {
+        match self {
+            SyntheticDisplayRotation::NoRotation => screen_size,
+            SyntheticDisplayRotation::Rotate90 => {
+                PhysicalSize::new(screen_size.height, screen_size.width)
+            }
+            SyntheticDisplayRotation::Rotate180 => screen_size,
+            SyntheticDisplayRotation::Rotate270 => {
+                PhysicalSize::new(screen_size.height, screen_size.width)
+            }
+        }
+    }
+
+    pub fn degrees(&self) -> f32 {
+        match self {
+            SyntheticDisplayRotation::NoRotation => 0.,
+            SyntheticDisplayRotation::Rotate90 => 90.,
+            SyntheticDisplayRotation::Rotate180 => 180.,
+            SyntheticDisplayRotation::Rotate270 => 270.,
+        }
+    }
+
+    pub fn translation_after_rotation(&self, screen_size: PhysicalSize) -> (f32, f32) {
+        match self {
+            SyntheticDisplayRotation::NoRotation => (0., 0.),
+            SyntheticDisplayRotation::Rotate90 => (0., -(screen_size.width as f32)),
+            SyntheticDisplayRotation::Rotate180 => {
+                (-(screen_size.width as f32), -(screen_size.height as f32))
+            }
+            SyntheticDisplayRotation::Rotate270 => (-(screen_size.height as f32), 0.),
+        }
+    }
+}